// packages/core/src/state/sessions.gen.ts
export enum SignatureType {
  EIP712 = 'EIP712',
  EthSign = 'EthSign',
  EIP1271 = 'EIP1271',
}
// ...
export interface Context {
  version: number
  factory: string
  mainModule: string
  mainModuleUpgradable: string
  guestModule: string
  walletCreationCode: string
}

export interface Signature {
  digest: string
  toImageHash?: string
  chainID: string
  type: SignatureType
  signature: string
  validOnChain?: string
  validOnBlock?: string
  validOnBlockHash?: string
}

export interface SignerSignature {
  signer?: string
  signature: string
  referenceChainID?: string
}

export interface ConfigUpdate {
  toImageHash: string
  signature: string
}

export interface Transaction {
  to: string
  value?: string
  data?: string
  gasLimit?: string
  delegateCall?: boolean
  revertOnError?: boolean
}

export interface TransactionBundle {
  executor: string
  transactions: Array<Transaction>
  nonce: string
  signature: string
}

export interface Sessions {
  ping(headers?: object, signal?: AbortSignal): Promise<PingReturn>
  config(args: ConfigArgs, headers?: object, signal?: AbortSignal): Promise<ConfigReturn>
  wallets(args: WalletsArgs, headers?: object, signal?: AbortSignal): Promise<WalletsReturn>
  deployHash(args: DeployHashArgs, headers?: object, signal?: AbortSignal): Promise<DeployHashReturn>
  configUpdates(args: ConfigUpdatesArgs, headers?: object, signal?: AbortSignal): Promise<ConfigUpdatesReturn>
  migrations(args: MigrationsArgs, headers?: object, signal?: AbortSignal): Promise<MigrationsReturn>
  saveConfig(args: SaveConfigArgs, headers?: object, signal?: AbortSignal): Promise<SaveConfigReturn>
  saveWallet(args: SaveWalletArgs, headers?: object, signal?: AbortSignal): Promise<SaveWalletReturn>
  saveSignature(args: SaveSignatureArgs, headers?: object, signal?: AbortSignal): Promise<SaveSignatureReturn>
  saveSignerSignatures(
    args: SaveSignerSignaturesArgs,
    headers?: object,
    signal?: AbortSignal,
  ): Promise<SaveSignerSignaturesReturn>
  saveSignerSignatures2(
    args: SaveSignerSignatures2Args,
    headers?: object,
    signal?: AbortSignal,
  ): Promise<SaveSignerSignatures2Return>
  saveMigration(args: SaveMigrationArgs, headers?: object, signal?: AbortSignal): Promise<SaveMigrationReturn>
}

export interface PingArgs {}

export interface PingReturn {}
export interface ConfigArgs {
  imageHash: string
}

export interface ConfigReturn {
  version: number
  config: any
}
export interface WalletsArgs {
  signer: string
  cursor?: number
  limit?: number
}

export interface WalletsReturn {
  wallets: { [key: string]: Signature }
  cursor: number
}
export interface DeployHashArgs {
  wallet: string
}

export interface DeployHashReturn {
  deployHash: string
  context: Context
}
export interface ConfigUpdatesArgs {
  wallet: string
  fromImageHash: string
  allUpdates?: boolean
}

export interface ConfigUpdatesReturn {
  updates: Array<ConfigUpdate>
}
export interface MigrationsArgs {
  wallet: string
  fromVersion: number
  fromImageHash: string
  chainID?: string
}

export interface MigrationsReturn {
  migrations: {
    [key: string]: { [key: number]: { [key: string]: TransactionBundle } }
  }
}
export interface SaveConfigArgs {
  version: number
  config: any
}

export interface SaveConfigReturn {}
export interface SaveWalletArgs {
  version: number
  deployConfig: any
}

export interface SaveWalletReturn {}
export interface SaveSignatureArgs {
  wallet: string
  digest: string
  chainID: string
  signature: string
  toConfig?: any
}

export interface SaveSignatureReturn {}
export interface SaveSignerSignaturesArgs {
  wallet: string
  digest: string
  chainID: string
  signatures: Array<string>
  toConfig?: any
}

export interface SaveSignerSignaturesReturn {}
export interface SaveSignerSignatures2Args {
  wallet: string
  digest: string
  chainID: string
  signatures: Array<SignerSignature>
  toConfig?: any
}

export interface SaveSignerSignatures2Return {}
export interface SaveMigrationArgs {
  wallet: string
  fromVersion: number
  toVersion: number
  toConfig: any
  executor: string
  transactions: Array<Transaction>
  nonce: string
  signature: string
  chainID?: string
}

export interface SaveMigrationReturn {}